+++ packages/reference/src/__fixtures__/FakeSdk.ts
@@ -1,19 +1,14 @@
-import { ContentType, FieldAPI, Link } from '@contentful/app-sdk';
+import { ContentType, FieldAPI, FieldExtensionSDK, Link } from '@contentful/app-sdk';
 import {
   createFakeCMAAdapter,
   createFakeFieldAPI,
   createFakeLocalesAPI,
   createFakeSpaceAPI,
 } from '@contentful/field-editor-test-utils';
+import { CollectionProp, GetSpaceParams, LocaleProps } from 'contentful-management';
 import { Emitter } from 'mitt';
 
-import { FieldExtensionSDK } from '../types';
-import changedAsset from './changed_asset.json';
-import changedEntry from './changed_entry.json';
-import emptyAsset from './empty_asset.json';
-import emptyEntry from './empty_entry.json';
-import publishedAsset from './published_asset.json';
-import publishedEntry from './published_entry.json';
+import { assets, contentTypes, entries, locales as localesFixtures, spaces } from './fixtures';
 
 const newLink = (linkType: string, id: string): Link => ({
   sys: {
@@ -44,14 +39,14 @@ export function newReferenceEditorFakeSdk(
   const space = createFakeSpaceAPI();
   const locales = createFakeLocalesAPI();
   const entryLinks = [
-    newLink('Entry', publishedEntry.sys.id),
-    newLink('Entry', changedEntry.sys.id),
-    newLink('Entry', emptyEntry.sys.id),
+    newLink('Entry', entries.published.sys.id),
+    newLink('Entry', entries.changed.sys.id),
+    newLink('Entry', entries.empty.sys.id),
   ];
   const assetLinks = [
-    newLink('Asset', publishedAsset.sys.id),
-    newLink('Asset', changedAsset.sys.id),
-    newLink('Asset', emptyAsset.sys.id),
+    newLink('Asset', assets.published.sys.id),
+    newLink('Asset', assets.changed.sys.id),
+    newLink('Asset', assets.empty.sys.id),
   ];
   let selectorCounter = 0;
 
@@ -78,16 +73,16 @@ export function newReferenceEditorFakeSdk(
           if (props?.fetchDelay) {
             await delay(props.fetchDelay);
           }
-          if (entryId === emptyEntry.sys.id) {
-            return emptyEntry;
+          if (entryId === entries.empty.sys.id) {
+            return entries.empty;
           }
-          if (entryId === publishedEntry.sys.id) {
-            return publishedEntry;
+          if (entryId === entries.published.sys.id) {
+            return entries.published;
           }
-          if (entryId === changedEntry.sys.id) {
-            return changedEntry;
+          if (entryId === entries.changed.sys.id) {
+            return entries.changed;
           }
-          return Promise.reject();
+          return Promise.reject({});
         },
       },
       Asset: {
@@ -95,17 +90,36 @@ export function newReferenceEditorFakeSdk(
           if (props?.fetchDelay) {
             await delay(props.fetchDelay);
           }
-          if (assetId === emptyAsset.sys.id) {
-            return emptyAsset;
+          if (assetId === assets.empty.sys.id) {
+            return assets.empty;
           }
-          if (assetId === publishedAsset.sys.id) {
-            return publishedAsset;
+          if (assetId === assets.published.sys.id) {
+            return assets.published;
           }
-          if (assetId === changedAsset.sys.id) {
-            return changedAsset;
+          if (assetId === assets.changed.sys.id) {
+            return assets.changed;
           }
-          return Promise.reject();
+          return Promise.reject({});
         },
+      },
+      Space: {
+        get: async (params: GetSpaceParams) => {
+          if (params.spaceId === spaces.indifferent.sys.id) {
+            return spaces.indifferent;
+          }
+          return Promise.reject({});
+        },
+      },
+      ContentType: {
+        get: async ({ contentTypeId }) => {
+          if (contentTypeId === contentTypes.published.sys.id) {
+            return contentTypes.published;
+          }
+          return Promise.reject({});
+        },
+      },
+      Locale: {
+        getMany: async () => localesFixtures.list as CollectionProp<LocaleProps>,
       },
     }),
     space: {
@@ -147,14 +161,14 @@ export function newReferenceEditorFakeSdk(
     },
     navigator: {
       openNewAsset: async () => ({
-        entity: newLink('Asset', emptyAsset.sys.id),
+        entity: newLink('Asset', assets.empty.sys.id),
       }),
       openAsset: async () => {
         alert('open Asset in slide in');
         return {};
       },
       openNewEntry: async () => ({
-        entity: newLink('Entry', emptyEntry.sys.id),
+        entity: newLink('Entry', entries.empty.sys.id),
       }),
       openEntry: async () => {
         alert('open entry in slide in');